# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Contract Management System API
  description: |
    Complete API documentation for the Contract Management System.
    
    ## Authentication
    All API requests require authentication using JWT tokens.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - Anonymous requests: 20/hour
    - Authenticated requests: 1000/hour
    - Premium users: 5000/hour
    
    ## WebSocket Support
    Real-time updates are available via WebSocket connections at:
    - Production: wss://api.contractmanagement.com/ws
    - Development: ws://localhost:8000/ws
    
  version: 1.0.0
  contact:
    name: API Support
    email: api@contractmanagement.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.contractmanagement.com/v1
    description: Production server
  - url: https://staging-api.contractmanagement.com/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Development server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Activity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entityType:
          type: string
          enum: [contract, template, user]
        entityId:
          type: string
          format: uuid
        action:
          type: string
        description:
          type: string
        userId:
          type: string
          format: uuid
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /analytics/dashboard:
    get:
      tags:
        - Analytics
      summary: Get dashboard analytics
      operationId: getDashboardAnalytics
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        200:
          description: Analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalContracts:
                    type: integer
                  activeContracts:
                    type: integer
                  totalValue:
                    type: number
                  contractsByStatus:
                    type: object
                  contractsByType:
                    type: object
                  recentActivity:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
  /analytics/contracts/{id}:
    get:
      tags:
        - Analytics
      summary: Get contract analytics
      operationId: getContractAnalytics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Contract analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  views:
                    type: integer
                  edits:
                    type: integer
                  comments:
                    type: integer
                  approvalTime:
                    type: number
                    description: Average approval time in hours
                  activity:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
...